Update HAOS upgrade path to include HAOS 9
Make sure the Supervisor knows the last HAOS 9 version. This should
avoid that people get 10.x offered directly, instead it guides people
to update to 9.5 first.

Note that there is no known issue to upgrade from any 9.x to 10
directly, so this shouldn't really matter in practice. But it is the
best tested and recommended path.