Use os-artifacts URL for stable's OTA updates too (#389)
Use the same URL for OTA (= operating system's RAUC update bundles) as
used on beta and dev. This will allow for updating also instances that
are configured to use the stable channel to update to dev and beta
versions if such version is explicitly requested. Currently it's
possible the other way, but without changing at least to the beta
channel, it's not possible to use beta/dev (e.g. for regression
testing).