Add HA Yellow Image to RPi imager index
Add raw HA Yellow image to list of options available for CM4/CM5
(although CM5 is yet not listed in the menu, it's presumed it will be
common with Pi 5 like it's for CM4).

The image can be used when the device is booted using rpiboot (which is
at this time the only option for installing the image to eMMC on CM5) or
when the installation should run from an NVMe drive. Adjusted image
descriptions to be more clear about the two Yellow options.